Added confirm permissions page

+++ src/SFA.DAS.ProviderRelationships.ReadStore/Application/Commands/UpdateRelationshipCommand.cs
@@ -8,14 +8,14 @@ namespace SFA.DAS.ProviderRelationships.ReadStore.Application.Commands
     public class UpdateRelationshipCommand : IReadStoreRequest<Unit>
     {
         public long Ukprn { get; }
-        public int AccountProviderId { get; }
+        public long AccountProviderId { get; }
         public long AccountId { get; }
         public long AccountLegalEntityId { get; }
         public HashSet<Operation> Operations { get; }
         public string MessageId { get; }
         public DateTime Updated { get; }
 
-        public UpdateRelationshipCommand(long ukprn, int accountProviderId, long accountId,long accountLegalEntityId, 
+        public UpdateRelationshipCommand(long ukprn, long accountProviderId, long accountId,long accountLegalEntityId, 
             HashSet<Operation> grants, string messageId, DateTime updated)
         {
             Ukprn = ukprn;
